Fixed elevators (again)
As titled. I have fixed the elevators again introducing a new penalty system. Before, we use to remove elevators from graph when function call was for a path on the same floor. However, now, any edge leading to an elevator has a penalty weight to discourage the algorithm, as well as an elevator switch penalty so that when engaged on a path the algorithm sticks to the same elevator shaft.